requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
yt-dlp==2024.3.10
faster-whisper==1.0.3
ffmpeg-python==0.2.0
deep-translator==1.11.4
python-multipart==0.0.9
jinja2==3.1.3
torch==2.2.2